+++ commands/bootstrap/inputs.go
@@ -17,11 +17,13 @@ package bootstrap
 import (
 	"fmt"
 	"io/ioutil"
+	"log"
 	"os"
 	"os/exec"
 	"path/filepath"
 	"strconv"
 	"strings"
+	"time"
 
 	"github.com/ystia/yorc/helper/collections"
 	"github.com/ystia/yorc/rest"
@@ -33,8 +35,8 @@ import (
 	"github.com/ystia/yorc/commands"
 	"github.com/ystia/yorc/config"
 
-	"gopkg.in/AlecAivazis/survey.v1"
-	"gopkg.in/yaml.v2"
+	survey "gopkg.in/AlecAivazis/survey.v1"
+	yaml "gopkg.in/yaml.v2"
 )
 
 var inputValues TopologyValues
@@ -331,9 +333,32 @@ func initializeInputs(inputFilePath, resourcesPath string, configuration config.
 		return err
 	}
 
+	//generating a deployment name if needed.
+	if deploymentID == "" {
+		t := time.Now()
+		deploymentID = fmt.Sprintf("bootstrap-%d-%02d-%02d--%02d-%02d-%02d",
+			t.Year(), t.Month(), t.Day(), t.Hour(), t.Minute(), t.Second())
+	}
+
 	// Get infrastructure from viper configuration if not provided in inputs
 	if inputValues.Infrastructures == nil {
 		inputValues.Infrastructures = configuration.Infrastructures
+	}
+
+	//recovering infra type if needed
+	if infrastructureType == "" {
+		switch inputValues.Location.Type {
+		case "OpenStack":
+			infrastructureType = "openstack"
+		case "Google Cloud":
+			infrastructureType = "google"
+		case "AWS":
+			infrastructureType = "aws"
+		case "HostsPool":
+			infrastructureType = "hostspool"
+		default:
+			infrastructureType = ""
+		}
 	}
 	// Now check for missing mandatory parameters and ask them to the user
 
@@ -596,7 +621,14 @@ func initializeInputs(inputFilePath, resourcesPath string, configuration config.
 		}
 	}
 
+	exportInputs()
+
+	if configOnly == true {
+		println("config_only option is set, exiting.")
+		os.Exit(0)
+	}
 	return nil
+
 }
 
 // reviewAndUpdateInputs allows the user to review and change deployment inputs
@@ -646,6 +678,51 @@ func reviewAndUpdateInputs() error {
 	err = yaml.Unmarshal([]byte(reply), &inputValues)
 
 	return nil
+}
+
+func exportInputs() {
+	if inputsPath == "" {
+
+		bSlice, err := yaml.Marshal(inputValues)
+
+		if err != nil {
+			log.Fatal("cannot marshal inputValues ", err)
+		}
+
+		inputsPathOut := deploymentID + ".yaml"
+
+		count := 1
+		for {
+			if _, err := os.Stat(inputsPathOut); os.IsNotExist(err) {
+				break
+			}
+			inputsPathOut = deploymentID + "_" + strconv.Itoa(count) + ".yaml"
+			count++
+		}
+
+		if _, err := os.Stat("inputsaves/"); os.IsNotExist(err) {
+			err = os.Mkdir("inputsaves/", 0700)
+			if err != nil {
+				log.Fatal("cannot create inputsave directory ", err)
+			}
+		}
+
+		println("Exporting set configuration to inputsaves/" + inputsPathOut)
+
+		file, err := os.OpenFile("inputsaves/"+inputsPathOut, os.O_CREATE|os.O_WRONLY, 0600)
+
+		if err != nil {
+			log.Fatal("Cannot create file to save the configuration ", err)
+		}
+
+		_, err = fmt.Fprintf(file, string(bSlice[:]))
+
+		if err != nil {
+			log.Fatal("Cannot write configuration to file", err)
+		}
+
+		file.Close()
+	}
 }
 
 // getComputeCredentials asks for the user used to connect to an on-demand compute node
@@ -783,11 +860,38 @@ func getHostsInputs(resourcesPath string) ([]rest.HostConfig, error) {
 			hostConfig.Connection.Port = 22
 		}
 
-		fmt.Println("Defining key/value labels for this host, a public_address label should be defined")
+		//asking public address first
+		fmt.Println("Defining key/value labels for this host. Defining public_address value is mandatory")
 		hostConfig.Labels = make(map[string]string)
-		labelsFinished := false
-		publicAddressDefined := false
-		for !labelsFinished {
+
+		prompt = &survey.Input{
+			Message: "public_address value:"}
+
+		question = &survey.Question{
+			Name:     "value",
+			Prompt:   prompt,
+			Validate: survey.Required,
+		}
+		if err := survey.Ask([]*survey.Question{question}, &answer); err != nil {
+			return nil, err
+		}
+
+		hostConfig.Labels["public_address"] = answer.Value
+
+		//asking for additional key values labels
+		for {
+
+			// asking if a new label must be defined
+			promptEnd := &survey.Select{
+				Message: "Add another key/value label for this host ?:",
+				Options: []string{"yes", "no"},
+				Default: "no",
+			}
+			var reply string
+			survey.AskOne(promptEnd, &reply, nil)
+			if reply == "no" {
+				break
+			}
 
 			prompt := &survey.Input{
 				Message: "Label key (required):"}
@@ -802,10 +906,6 @@ func getHostsInputs(resourcesPath string) ([]rest.HostConfig, error) {
 			}
 			labelKey := answer.Value
 
-			if labelKey == "public_address" {
-				publicAddressDefined = true
-			}
-
 			prompt = &survey.Input{
 				Message: "Label value:"}
 
@@ -818,22 +918,6 @@ func getHostsInputs(resourcesPath string) ([]rest.HostConfig, error) {
 			}
 
 			hostConfig.Labels[labelKey] = answer.Value
-
-			// If the public address label is already defined
-			// asking if a new label must be defined
-			// else a new label must be defined, because public_address is
-			// mandatory for the bootstrap
-			if publicAddressDefined {
-				promptEnd := &survey.Select{
-					Message: "Add another key/value label:",
-					Options: []string{"yes", "no"},
-					Default: "no",
-				}
-				var reply string
-				survey.AskOne(promptEnd, &reply, nil)
-				labelsFinished = (reply == "no")
-			}
-
 		}
 
 		// The private SSH key used to connect to the host is the Yorc private key
@@ -1032,6 +1116,9 @@ func getInputValues(inputFilePath string) (TopologyValues, error) {
 				fmt.Println("Can't use config file:", err)
 			}
 		}
+
+		var infratype = bootstrapViper.GetString("InfrastructureType")
+		infrastructureType = infratype
 	}
 
 	err := bootstrapViper.Unmarshal(&values)